requirements.txt
#Build agents and workflows to automate intelligent work.

phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq

